wrote code for loading rem files (might be modified later to decrypt media files beforehand); started implementation for loading rem files UI integration (incomplete); fixed createRemFile bug

+++ lib/features/data_loader/data_loader.dart
@@ -1,21 +1,46 @@
 import 'dart:io';
+import 'dart:isolate';
 
 import 'package:archive/archive.dart';
 import 'package:path/path.dart' as p;
 import 'package:path_provider/path_provider.dart';
 import 'package:reminiscence/features/data_loader/utils.dart';
 
-import 'package:reminiscence/features/database/database.dart';
-import "package:reminiscence/features/data_loader/reminiscence_file.dart";
+import "package:reminiscence/features/data_loader/reminiscence_data.dart";
 
-Future<ReminiscenceFile?> loadRemFile(String filePath, String? password) async {
+Future<ReminiscenceData?> loadRemFile({
+  required String filePath,
+  String? password,
+  required SendPort? sendPort,
+}) async {
+  // Set up receive port and cancellation token
+  ReceivePort? receivePort = sendPort != null ? ReceivePort() : null;
+  bool isCancelled = false;
+
+  if (receivePort != null) {
+    sendPort!.send({"type": "sendPort", "sendPort": receivePort.sendPort});
+  }
+
+  receivePort?.listen((message) {
+    if (message is! Map<String, dynamic>) return;
+
+    if (message["type"] == "cancel") {
+      isCancelled = true;
+    }
+  });
+
+  // Prepare all the extracted paths
   final Directory tempDir = await getTemporaryDirectory();
-  final String databasePath = p.join(tempDir.path, "database.db");
+  final String dbPath = p.join(tempDir.path, "database.db");
   final Directory mediaDir = Directory(p.join(tempDir.path, "media"));
 
+  if (isCancelled) return null;
+
+  // Open the rem file
   InputFileStream stream = InputFileStream(filePath);
   Archive archive = ZipDecoder().decodeStream(stream);
 
+  // Get the database and nonce files from the rem file.
   ArchiveFile? databaseArchiveFile = archive.find("database.db");
   ArchiveFile? nonceArchiveFile = archive.find("nonce.txt");
 
@@ -24,9 +49,19 @@ Future<ReminiscenceFile?> loadRemFile(String filePath, String? password) async {
     return null;
   }
 
-  final dbStream = OutputFileStream(databasePath);
+  if (isCancelled) return null;
+
+  sendPort?.send({
+    "type": "progress",
+    "progress": {"value": 0.0, "label": 'Preparing Chat Messages...'},
+  });
+
+  // Extract the database file.
+  final dbStream = OutputFileStream(dbPath);
   databaseArchiveFile.writeContent(dbStream);
   await dbStream.close();
+
+  if (isCancelled) return null;
 
   // Get the nonce
   List<int> nonce = [];
@@ -35,12 +70,31 @@ Future<ReminiscenceFile?> loadRemFile(String filePath, String? password) async {
     nonce = nonceArchiveFile.readBytes() ?? [];
   }
 
+  if (isCancelled) return null;
+
+  sendPort?.send({
+    "type": "progress",
+    "progress": {"value": 0.5, "label": 'Preparing Chat Media...'},
+  });
+
   // Extract the media files
   await mediaDir.delete(recursive: true);
+
+  if (isCancelled) return null;
+
   await extractArchiveDir(archive, "media", mediaDir);
 
-  // Load the database
-  AppDatabase db = AppDatabase(dbPath: databasePath, password: password);
+  if (isCancelled) return null;
 
-  return ReminiscenceFile(db: db, nonce: nonce, mediaDir: mediaDir);
+  sendPort?.send({
+    "type": "progress",
+    "progress": {"value": 1.0},
+  });
+
+  return ReminiscenceData(
+    dbPath: dbPath,
+    password: password,
+    nonce: nonce,
+    mediaDir: mediaDir,
+  );
 }
